pyproject.toml
[project]
name = "django-ninja-api-books"
version = "0.1.0"
description = "Default template for PDM package"
authors = [{ name = "teigorme", email = "igorteixeira354@gmail.com" }]
dependencies = ["django-ninja>=1.4.5", "pydantic-settings>=2.11.0", "psycopg2>=2.9.11", "uvicorn>=0.38.0"]
requires-python = "==3.13.*"
readme = "README.md"
license = { text = "MIT" }


[tool.pdm]
distribution = false

[tool.pdm.scripts]
dev = "uvicorn config.asgi:application --reload"
